Trait tuna::builder::Projection

pub trait Projection<Src>: ToSql {
    type Value;
}

Defines the types of things that can be projected from a source Src (probably a table).

Associated Types

type Value

The type that the projected thing gets in Rust. See Column below for an example of where this comes from.

Loading content...