Trait tuna::builder::ToSql

pub trait ToSql {
    type Sql: Display;
    fn sql(&self) -> Self::Sql;
}

Types that can be used to generate SQL.

Associated Types

type Sql: Display

The result type. This is something that can be displayed, but not necessarily a string. This makes it possible to return structures with multiple types and display those parts in turn without allocating.

Loading content...
Required methods

fn sql(&self) -> Self::Sql

Returns the SQL for this thing.

Loading content...