Struct tuna::column::Column

pub struct Column<Table, Type> {
    pub name: sstr,
    pub position: usize,
    pub parse: fn'a(_: &'a str) -> Type,
    pub _table_marker: PhantomData<fn(_: &Table)>,
}

Represents some column of a table whose values convert to a Rust type. The column keeps track of the data source (table) through the use of PhantomData.

For example, suppose that there is a table type UserTable: Selectable, with these two fields:

id is a SQL INTEGER NOT NULL

login_count is a SQL INTEGER

Then you would want to define columns like so:

struct UserTable;

const ID: Column<UserTable, i64> = Column {
    name:     "id",
    position: 0,
    parse:    |s| s.parse().unwrap(), // parsing for now
    _marker:  PhantomData,
};

const LOGIN_COUNT: Column<UserTable, Option<u64>> = Column {
    name:     "login_count",
    position: 3,
    parse:    |s| s.parse().ok(), // parsing for now
    _marker:  PhantomData,
};

Fields
name: sstr

The name of the column.

position: usize

The zero based index of the column in the table.

parse: fn'a(_: &'a str) -> Type

How to data from this column into a Rust Type

_table_marker: PhantomData<fn(_: &Table)>

Keeps track of the source of data for this column

Methods

impl<Table, Type> Column<Table, Type> where
Table: Selectable,

pub fn equals(self, other: Type) -> Equals<Table, Self, Type>

Checks to see if the column has data that equals some other value.

pub fn null(self) -> IsNull<Table, Self>

Checks to see if the column has data that is null.

pub fn not_null(self) -> IsNotNull<Table, Self>

Checks to see if the column has data that is not null.

pub fn greater_than(self, other: Type) -> Greater<Table, Self, Type>

Checks to see if the column has data that is greater than some other value.

pub fn less_than(self, other: Type) -> Less<Table, Self, Type>

Checks to see if the column has data that is less than some other value.

pub fn leq(self, other: Type) -> Leq<Table, Self, Type>

Checks to see if the column has data that is less than or equal to some other value.

pub fn geq(self, other: Type) -> Geq<Table, Self, Type>

Checks to see if the column has data that is greater than or equal to some other value.

pub fn not_equals(self, other: Type) -> NotEq<Table, Self, Type>

Checks to see if the column has data that is not equal to some other value.