Trait tuna::condition::Condition

pub trait Condition<Src>: ToSql {
    fn and<Other: Condition<Src>>(self, other: Other) -> Both<Self, Other>
    where
        Self: Sized,
    { ... }

    fn or<Other: Condition<Src>>(self, other: Other) -> Either<Self, Other>
    where
        Self: Sized,
    { ... }

    fn not(self) -> Not<Self>
    where
        Self: Sized,
    { ... }
}

A trait to represent a condition to put in a SQL WHERE clause. The parameter is the source of the data for the condition, which prevents us from referring to columns that don't exist in the given query's source table.

Provided methods

fn and<Other: Condition<Src>>(self, other: Other) -> Both<Self, Other> where
Self: Sized,

Chains two conditions from the same souce together using the SQL AND clause.

fn or<Other: Condition<Src>>(self, other: Other) -> Either<Self, Other> where
Self: Sized,

Chains two conditions from the same source together using the SQL OR clause.

fn not(self) -> Not<Self> where
Self: Sized,

Negates a condition using the SQL NOT clause.

Loading content...